+++ CHANGELOG.md
@@ -26,6 +26,10 @@ following conventions:
 
 ## Unreleased
 
+### Non-Breaking
+
+* Bump `bytestring` dependency version upper bound
+
 ## 1.2.1.0 (2023-03-21)
 
 ### Non-Breaking